0xMirror: Bug Found
The destination path is being created by replacing src_root by dest_root
in cur_dir.

This works fine when both src and dest have an ending slash (\)

When they don't the destination path is not correct.

The proper way of doing this must be by using path.split or something
but would take time...